fix: 修复未生成config.json时缓存错误的Bug

+++ packages/uni-uts-v1/src/manifest/config.ts
@@ -31,7 +31,7 @@ export function restoreConfigJson(
     cacheDir,
     platform
   )
-  if (cacheFile) {
+  if (cacheFile && existsSync(cacheFile)) {
     copySync(
       cacheFile,
       join(outputDir, pluginRelativeDir, 'utssdk', platform, 'config.json')
